requirements.txt
aiogram==2.5
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
yarl==1.9.2
